support markdown description
